+++ resources/views/standings/race.blade.php
@@ -118,11 +118,11 @@ td {
             <thead>
               <tr>
               @auth
-                @view('admin,coordinator,steward')
+                @can('admin|coordinator|steward')
                 <th class="rounded-lg bg-gray-800 tracking-widest text-gray-100 border-2 border-white text-center w-1/12 confTable">D.ID</th>
                 <th class="rounded-lg bg-gray-800 tracking-widest text-gray-100 border-2 border-white text-center w-1/12 confTable">C.ID</th>
                   
-                @endview
+                @endcan
               @endauth    
                 
                 <th class="rounded-lg bg-gray-800 tracking-widest text-gray-100 border-2 border-white text-center w-1/12">Pos.</th>
@@ -136,11 +136,11 @@ td {
         @if((int)$results[$i]['status'] % 10 == 1)
           <tr>
             @auth
-              @view('admin,coordinator,steward')
+              @can('admin|coordinator|steward')
               <td class="font-semibold rounded-lg border border-white bg-purple-200 text-purple-700 text-center tracking-widest confTable">{{$results[$i]['driver']['id']}}</td>
               <td class="font-semibold rounded-lg border border-white bg-purple-200 text-purple-700 text-center tracking-widest confTable">{{$results[$i]['constructor_id']}}</td>
                 
-              @endview
+              @endcan
               @endauth    
                 
               <td class="font-semibold rounded-lg border border-white bg-purple-200 text-purple-700 text-center tracking-widest">{{$i+1}}</td>
@@ -170,7 +170,7 @@ td {
         @elseif($i % 2 != 0)
         <tr>
             @auth
-            @view('admin,coordinator,steward')
+            @can('admin|coordinator|steward')
               @if ($results[$i]['driver']['user_id'] == Auth::id())
                 <td class="font-semibold rounded-lg border border-white bg-gray-700 text-white text-center tracking-widest confTable">{{$results[$i]['driver']['id']}}</td>
               @else
@@ -182,7 +182,7 @@ td {
               @else
                 <td class="font-semibold rounded-lg border border-white bg-gray-200 text-center tracking-widest confTable">{{$results[$i]['constructor_id']}}</td>
               @endif    
-            @endview
+            @endcan
             @endauth    
             
             @if ($results[$i]['driver']['user_id'] == Auth::id())
@@ -259,7 +259,7 @@ td {
             </tr> -->
             <tr>
             @auth
-              @view('admin,coordinator,steward')
+              @can('admin|coordinator|steward')
                 @if ($results[$i]['driver']['user_id'] == Auth::id())
                   <td class="font-semibold rounded-lg border border-white bg-gray-700 text-white text-center tracking-widest confTable">{{$results[$i]['driver']['id']}}</td>
                 @else
@@ -272,7 +272,7 @@ td {
                   <td class="font-semibold rounded-lg border border-white text-center tracking-widest confTable">{{$results[$i]['constructor_id']}}</td>
                 @endif
               
-              @endview
+              @endcan
             @endauth    
         
             @if ($results[$i]['driver']['user_id'] == Auth::id())
@@ -345,7 +345,7 @@ td {
 <!-- Admin View -->
 <div>
 @auth
-   @view('admin,coordinator,steward')
+   @can('admin|coordinator|steward')
    <div class="border p-5 rounded-lg confTable">
       <div class="text-2xl font-bold text-center">Admin/Coordinator Information</div>
       <div class="flex flex-wrap gap-5">
@@ -361,12 +361,12 @@ td {
         
       </div>
    </div>
-   @endview
+   @endcan
 @endauth
 </div>
 
 @auth
-   @view('admin,coordinator,steward')
+   @can('admin|coordinator|steward')
     <script>
       $('body').keypress(function (e) { 
         if(e.keyCode == 72 || e.keyCode == 104){
@@ -375,7 +375,7 @@ td {
       });
       $('.confTable').toggle();
     </script>
-  @endview
+  @endcan
 @endauth
 
 @endsection